Feature Update
Now the IP is no longer hard coded, it is passed with the -s <IP OR
HOST> option. You can now use a host name instead of an IP.  By default
the program only checks the server then exits. If you want to keep it
open indefinitely pass the --watch flag.